Refactor move PageTreeReadApi into own file, this file got too large
